// src/shared/api/models/postApiSaleBodyData.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiSaleBodyDataCustomerId } from './postApiSaleBodyDataCustomerId';
import type { PostApiSaleBodyDataSalesType } from './postApiSaleBodyDataSalesType';
import type { PostApiSaleBodyDataItemsItem } from './postApiSaleBodyDataItemsItem';
import type { PostApiSaleBodyDataSalesTaxesItem } from './postApiSaleBodyDataSalesTaxesItem';
import type { PostApiSaleBodyDataSalesDiscountsItem } from './postApiSaleBodyDataSalesDiscountsItem';
import type { PostApiSaleBodyDataPaymentItem } from './postApiSaleBodyDataPaymentItem';

export type PostApiSaleBodyData = {
    salesDate?: string;
    customerId?: PostApiSaleBodyDataCustomerId;
    customerName?: string;
    subTotal: number;
    totalTax: number;
    netTotal: number;
    tenderAmount: number;
    change: number;
    roundOff: number;
    tips?: number;
    salesNote?: string;
    salesType?: PostApiSaleBodyDataSalesType;
    sellBy: number;
    items: PostApiSaleBodyDataItemsItem[];
    salesTaxes?: PostApiSaleBodyDataSalesTaxesItem[];
    salesDiscounts?: PostApiSaleBodyDataSalesDiscountsItem[];
    payment: PostApiSaleBodyDataPaymentItem[];
};