// src/shared/api/models/postApiSaleBodyDataPaymentItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiSaleBodyDataPaymentItemPaymentId } from './postApiSaleBodyDataPaymentItemPaymentId';
import type { PostApiSaleBodyDataPaymentItemTxStatus } from './postApiSaleBodyDataPaymentItemTxStatus';
import type { PostApiSaleBodyDataPaymentItemTerminalId } from './postApiSaleBodyDataPaymentItemTerminalId';
import type { PostApiSaleBodyDataPaymentItemTerminalRefId } from './postApiSaleBodyDataPaymentItemTerminalRefId';
import type { PostApiSaleBodyDataPaymentItemTerminalStatus } from './postApiSaleBodyDataPaymentItemTerminalStatus';
import type { PostApiSaleBodyDataPaymentItemCardType } from './postApiSaleBodyDataPaymentItemCardType';
import type { PostApiSaleBodyDataPaymentItemCardFourDigit } from './postApiSaleBodyDataPaymentItemCardFourDigit';
import type { PostApiSaleBodyDataPaymentItemPaymentType } from './postApiSaleBodyDataPaymentItemPaymentType';
import type { PostApiSaleBodyDataPaymentItemBunddleOffer } from './postApiSaleBodyDataPaymentItemBunddleOffer';

export type PostApiSaleBodyDataPaymentItem = {
    paymentId?: PostApiSaleBodyDataPaymentItemPaymentId;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    amount?: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    tenderAmount?: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    change?: number;
    txStatus?: PostApiSaleBodyDataPaymentItemTxStatus;
    terminalId?: PostApiSaleBodyDataPaymentItemTerminalId;
    terminalRefId?: PostApiSaleBodyDataPaymentItemTerminalRefId;
    terminalStatus?: PostApiSaleBodyDataPaymentItemTerminalStatus;
    cardType?: PostApiSaleBodyDataPaymentItemCardType;
    cardFourDigit?: PostApiSaleBodyDataPaymentItemCardFourDigit;
    paymentType?: PostApiSaleBodyDataPaymentItemPaymentType;
    terminalResponse?: string;
    currency?: string;
    giftCardId?: string;
    bunddleOffer?: PostApiSaleBodyDataPaymentItemBunddleOffer;
};